requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google-genai
pyyaml
redis
langfuse>=2,<3